Add link to navigate to www.messenger.com under the Go To Messenger prompt
I don't think there's enough awareness about the www.messenger.com website (I wasn't even aware of it before we started this out!), and this should present the almost-seamless jump for people still looking for to the messenger in-browser that would otherwise assume it's fully gone (it isn't!).

Some styling may be welcome, but I think a discrete hyperlink a few units under the button is fine (it's how the "Continue doing whatever you wanted to do without installing the thing we want" options are usually presented anyway).

(also I guess this is sort of a test for the PR autobuild thing? idk)